[Chore #161695595] Optimize Database, Product Model
