+++ storemanager/api/v2/database/database.py
@@ -3,6 +3,9 @@ import psycopg2
 from storemanager.api.v2.database.config import config
 from .queries import *
 
+conn = None
+result = None
+
 
 class Database:
 
@@ -24,35 +27,52 @@ class Database:
             CREATE_TABLE_SALE_ITEMS
         ]
 
-        conn = None
         for statement in create_tables_query:
-            try:
-                conn = DB.connect()
-                cur = conn.cursor()
-                cur.execute(statement)
-                conn.commit()
-                cur.close()
-            except (Exception, psycopg2.DatabaseError) as error:
-                print(error)
-            finally:
-                if conn is not None:
-                    conn.close()
+            execute_query([statement], "one_no_result")
 
     @classmethod
     def drop_tables(cls):
         print('Dropping Tables')
-        conn = None
-        try:
-            conn = DB.connect()
-            cur = conn.cursor()
-            cur.execute(DROP_ALL_TABLES)
-            conn.commit()
-            cur.close()
-        except (Exception, psycopg2.DatabaseError) as error:
-            print(error)
-        finally:
-            if conn is not None:
-                conn.close()
+        execute_query([DROP_ALL_TABLES], "one_no_result")
+
+
+def execute_query(query, flag):
+    """Execute queries based on flag values"""
+    try:
+        global conn, result
+        statement = query[0]
+        conn = DB.connect()
+        cur = conn.cursor()
+
+        if flag is "one":
+            values = query[1]
+            cur.execute(statement, values)
+            result = cur.fetchone()
+
+        elif flag is "one_no_result":
+            cur.execute(statement)
+        elif flag is "many":
+            values = query[1]
+            cur.execute(statement, values)
+            result = cur.fetchall()
+
+        elif flag is "many_no_values":
+            cur.execute(statement)
+            result = cur.fetchall()
+
+        elif flag is "one_row_count":
+            value = query[1]
+            cur.execute(statement, value)
+            result = cur.rowcount
+        conn.commit()
+        cur.close()
+    except (Exception, psycopg2.DatabaseError) as error:
+        print(error)
+    finally:
+        if conn is not None:
+            conn.close()
+
+    return result
 
 
 DB = Database()
