+++ storemanager/api/v2/models/abstract_model.py
@@ -2,11 +2,7 @@
 This module contains the Generic Model
 UserModel, ProductModel and SaleRecordModel inherit from AbstractModel
 """
-import psycopg2
-from storemanager.api.v2.database.database import DB
-
-conn = None
-result = None
+from storemanager.api.v2.database.database import execute_query
 
 
 class AbstractModel:
@@ -51,40 +47,3 @@ class AbstractModel:
     def get_by_name(cls, statement, value):
         """Returns the entity with the specified name"""
         return execute_query([statement, value], "one")
-
-
-def execute_query(query, flag):
-    """Execute queries based on flag values"""
-    try:
-        global conn, result
-        statement = query[0]
-        conn = DB.connect()
-        cur = conn.cursor()
-
-        if flag is "one":
-            values = query[1]
-            cur.execute(statement, values)
-            result = cur.fetchone()
-
-        elif flag is "many":
-            values = query[1]
-            cur.execute(statement, values)
-            result = cur.fetchall()
-
-        elif flag is "many_no_values":
-            cur.execute(statement)
-            result = cur.fetchall()
-
-        elif flag is "one_row_count":
-            value = query[1]
-            cur.execute(statement, value)
-            result = cur.rowcount
-        conn.commit()
-        cur.close()
-    except (Exception, psycopg2.DatabaseError) as error:
-        print(error)
-    finally:
-        if conn is not None:
-            conn.close()
-
-    return result
